# docs/swagger/swagger.yaml
...
openapi: 3.0.1
info:
  title: WineSales
  description: The API for the WineSales project
  version: v1
paths:
  "/api/v1/customers":
    get:
      tags:
      - Customer
      responses:
        '200':
          description: Success
          content:
            text/plain:
              schema:
                type: array
                items:
                  "$ref": "#/components/schemas/CustomerDTO"
            application/json:
              schema:
                type: array
                items:
                  "$ref": "#/components/schemas/CustomerDTO"
            text/json:
              schema:
                type: array
                items:
                  "$ref": "#/components/schemas/CustomerDTO"
    post:
      tags:
      - Customer
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/CustomerDTO"
          text/json:
            schema:
              "$ref": "#/components/schemas/CustomerDTO"
          application/*+json:
            schema:
              "$ref": "#/components/schemas/CustomerDTO"
      responses:
        '201':
          description: Success
          content:
            text/plain:
              schema:
                "$ref": "#/components/schemas/CustomerDTO"
            application/json:
              schema:
                "$ref": "#/components/schemas/CustomerDTO"
            text/json:
              schema:
                "$ref": "#/components/schemas/CustomerDTO"
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '409':
          description: Conflict
  "/api/v1/customers/{id}":
    patch:
      tags:
      - Customer
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/CustomerBaseDTO"
          text/json:
            schema:
              "$ref": "#/components/schemas/CustomerBaseDTO"
          application/*+json:
            schema:
              "$ref": "#/components/schemas/CustomerBaseDTO"
      responses:
        '200':
          description: Success
          content:
            text/plain:
              schema:
                "$ref": "#/components/schemas/CustomerDTO"
            application/json:
              schema:
                "$ref": "#/components/schemas/CustomerDTO"
            text/json:
              schema:
                "$ref": "#/components/schemas/CustomerDTO"
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '404':
          description: Not Found
        '409':
          description: Conflict
    delete:
      tags:
      - Customer
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        '200':
          description: Success
          content:
            text/plain:
              schema:
                "$ref": "#/components/schemas/CustomerDTO"
            application/json:
              schema:
                "$ref": "#/components/schemas/CustomerDTO"
            text/json:
              schema:
                "$ref": "#/components/schemas/CustomerDTO"
        '401':
          description: Unauthorized
        '404':
          description: Not Found
    get:
      tags:
      - Customer
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        '200':
          description: Success
          content:
            text/plain:
              schema:
                "$ref": "#/components/schemas/CustomerDTO"
            application/json:
              schema:
                "$ref": "#/components/schemas/CustomerDTO"
            text/json:
              schema:
                "$ref": "#/components/schemas/CustomerDTO"
        '404':
          description: Not Found
...
  "/api/v1/suppliers/{id}":
    patch:
      tags:
      - Supplier
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/SupplierBaseDTO"
          text/json:
            schema:
              "$ref": "#/components/schemas/SupplierBaseDTO"
          application/*+json:
            schema:
              "$ref": "#/components/schemas/SupplierBaseDTO"
      responses:
        '200':
          description: Success
          content:
            text/plain:
              schema:
                "$ref": "#/components/schemas/SupplierDTO"
            application/json:
              schema:
                "$ref": "#/components/schemas/SupplierDTO"
            text/json:
              schema:
                "$ref": "#/components/schemas/SupplierDTO"
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '404':
          description: Not Found
        '409':
          description: Conflict
    delete:
      tags:
      - Supplier
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        '200':
          description: Success
          content:
            text/plain:
              schema:
                "$ref": "#/components/schemas/SupplierDTO"
            application/json:
              schema:
                "$ref": "#/components/schemas/SupplierDTO"
            text/json:
              schema:
                "$ref": "#/components/schemas/SupplierDTO"
        '401':
          description: Unauthorized
        '404':
          description: Not Found
    get:
      tags:
      - Supplier
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        '200':
          description: Success
          content:
            text/plain:
              schema:
                "$ref": "#/components/schemas/SupplierDTO"
            application/json:
              schema:
                "$ref": "#/components/schemas/SupplierDTO"
            text/json:
              schema:
                "$ref": "#/components/schemas/SupplierDTO"
        '404':
          description: Not Found
...
  "/api/v1/supplierWine":
    get:
      tags:
      - SupplierWine
      responses:
        '200':
          description: Success
          content:
            text/plain:
              schema:
                type: array
                items:
                  "$ref": "#/components/schemas/SupplierWineDTO"
            application/json:
              schema:
                type: array
                items:
                  "$ref": "#/components/schemas/SupplierWineDTO"
            text/json:
              schema:
                type: array
                items:
                  "$ref": "#/components/schemas/SupplierWineDTO"
    post:
      tags:
      - SupplierWine
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/SupplierWineDTO"
          text/json:
            schema:
              "$ref": "#/components/schemas/SupplierWineDTO"
          application/*+json:
            schema:
              "$ref": "#/components/schemas/SupplierWineDTO"
      responses:
        '201':
          description: Success
          content:
            text/plain:
              schema:
                "$ref": "#/components/schemas/SupplierWineDTO"
            application/json:
              schema:
                "$ref": "#/components/schemas/SupplierWineDTO"
            text/json:
              schema:
                "$ref": "#/components/schemas/SupplierWineDTO"
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '409':
          description: Conflict
  "/api/v1/supplierWine/{id}":
    patch:
      tags:
      - SupplierWine
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/SupplierWineBaseDTO"
          text/json:
            schema:
              "$ref": "#/components/schemas/SupplierWineBaseDTO"
          application/*+json:
            schema:
              "$ref": "#/components/schemas/SupplierWineBaseDTO"
      responses:
        '200':
          description: Success
          content:
            text/plain:
              schema:
                "$ref": "#/components/schemas/SupplierWineDTO"
            application/json:
              schema:
                "$ref": "#/components/schemas/SupplierWineDTO"
            text/json:
              schema:
                "$ref": "#/components/schemas/SupplierWineDTO"
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '404':
          description: Not Found
        '409':
          description: Conflict
    delete:
      tags:
      - SupplierWine
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        '200':
          description: Success
          content:
            text/plain:
              schema:
                "$ref": "#/components/schemas/SupplierWineDTO"
            application/json:
              schema:
                "$ref": "#/components/schemas/SupplierWineDTO"
            text/json:
              schema:
                "$ref": "#/components/schemas/SupplierWineDTO"
        '401':
          description: Unauthorized
        '404':
          description: Not Found
    get:
      tags:
      - SupplierWine
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        '200':
          description: Success
          content:
            text/plain:
              schema:
                "$ref": "#/components/schemas/SupplierWineDTO"
            application/json:
              schema:
                "$ref": "#/components/schemas/SupplierWineDTO"
            text/json:
              schema:
                "$ref": "#/components/schemas/SupplierWineDTO"
        '404':
          description: Not Found
  "/api/v1/supplierWine/{supplierWineId}/supplier":
    get:
      tags:
      - SupplierWine
      parameters:
      - name: id
        in: query
        schema:
          type: integer
          format: int32
      - name: supplierWineId
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Success
          content:
            text/plain:
              schema:
                "$ref": "#/components/schemas/SupplierDTO"
            application/json:
              schema:
                "$ref": "#/components/schemas/SupplierDTO"
            text/json:
              schema:
                "$ref": "#/components/schemas/SupplierDTO"
        '401':
          description: Unauthorized
        '404':
          description: Not Found
...
  "/api/v1/wines/{id}":
    patch:
      tags:
      - Wine
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/WineBaseDTO"
          text/json:
            schema:
              "$ref": "#/components/schemas/WineBaseDTO"
          application/*+json:
            schema:
              "$ref": "#/components/schemas/WineBaseDTO"
      responses:
        '200':
          description: Success
          content:
            text/plain:
              schema:
                "$ref": "#/components/schemas/WineDTO"
            application/json:
              schema:
                "$ref": "#/components/schemas/WineDTO"
            text/json:
              schema:
                "$ref": "#/components/schemas/WineDTO"
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '404':
          description: Not Found
        '409':
          description: Conflict
    delete:
      tags:
      - Wine
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        '200':
          description: Success
          content:
            text/plain:
              schema:
                "$ref": "#/components/schemas/WineDTO"
            application/json:
              schema:
                "$ref": "#/components/schemas/WineDTO"
            text/json:
              schema:
                "$ref": "#/components/schemas/WineDTO"
        '401':
          description: Unauthorized
        '404':
          description: Not Found
    get:
      tags:
      - Wine
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        '200':
          description: Success
          content:
            text/plain:
              schema:
                "$ref": "#/components/schemas/WineDTO"
            application/json:
              schema:
                "$ref": "#/components/schemas/WineDTO"
            text/json:
              schema:
                "$ref": "#/components/schemas/WineDTO"
        '401':
          description: Unauthorized
        '404':
          description: Not Found
components:
  schemas:
    CustomerBaseDTO:
      type: object
      properties:
        name:
          type: string
          nullable: true
        surname:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
      additionalProperties: false
    CustomerDTO:
      type: object
      properties:
        name:
          type: string
          nullable: true
        surname:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        id:
          type: integer
          format: int32
      additionalProperties: false
...
    SupplierBaseDTO:
      type: object
      properties:
        name:
          type: string
          nullable: true
        country:
          type: string
          nullable: true
        license:
          type: boolean
          nullable: true
      additionalProperties: false
    SupplierDTO:
      type: object
      properties:
        name:
          type: string
          nullable: true
        country:
          type: string
          nullable: true
        license:
          type: boolean
          nullable: true
        id:
          type: integer
          format: int32
      additionalProperties: false
    SupplierWineBaseDTO:
      type: object
      properties:
        supplierID:
          type: integer
          format: int32
          nullable: true
        wineID:
          type: integer
          format: int32
          nullable: true
        price:
          type: number
          format: double
          nullable: true
        percent:
          type: integer
          format: int32
          nullable: true
      additionalProperties: false
    SupplierWineDTO:
      type: object
      properties:
        supplierID:
          type: integer
          format: int32
          nullable: true
        wineID:
          type: integer
          format: int32
          nullable: true
        price:
          type: number
          format: double
          nullable: true
        percent:
          type: integer
          format: int32
          nullable: true
        id:
          type: integer
          format: int32
      additionalProperties: false
...
    WineBaseDTO:
      type: object
      properties:
        kind:
          type: string
          nullable: true
        color:
          type: string
          nullable: true
        sugar:
          type: string
          nullable: true
        volume:
          type: number
          format: double
          nullable: true
        alcohol:
          type: number
          format: double
          nullable: true
        number:
          type: integer
          format: int32
          nullable: true
      additionalProperties: false
    WineDTO:
      type: object
      properties:
        kind:
          type: string
          nullable: true
        color:
          type: string
          nullable: true
        sugar:
          type: string
          nullable: true
        volume:
          type: number
          format: double
          nullable: true
        alcohol:
          type: number
          format: double
          nullable: true
        number:
          type: integer
          format: int32
          nullable: true
        id:
          type: integer
          format: int32
      additionalProperties: false
  securitySchemes:
    Bearer:
      type: apiKey
      description: Please insert JWT with Bearer into field
      name: Authorization
      in: header
security:
- Bearer: []